requirements.txt
fastapi
uvicorn
torch
torchvision
albumentations
Pillow
python-multipart
